[IntraNodeComm] refactor rendezvous to allow for lazy initialization
We don't require users to set cuda device prior to initializing NCCL process
group, which initializes it communicators lazily. In order to make IntraNodeComm
an optional fast path for NCCL, we need to make it behave the same.

This PR refactors the handshaking (P2P buffer initialization) logic into
a separate method, so that the upcoming P2P `ProcessGroup` can lazily
initialize `IntraNodeComm`.

ghstack-source-id: 20bc3514887c9b8a9278fdbbd960167d81cd429d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120968