Consult un-simplified expressions when determining binding sites for runtime asserts
Obsoletes https://github.com/pytorch/pytorch/pull/120816 as it is much
simpler.

Fixes https://github.com/pytorch/pytorch/issues/119689

The idea per
https://github.com/pytorch/pytorch/pull/120816#discussion_r1509689819 is
to avoid losing track of when unbacked SymInts are allocated, is to make
sure that we look at _expr instead of expr to get the original symbol,
pre-replacement.

This strategy won't work later, when we have renamed unbacked SymInts
because we accidentally reallocated them due to retracing, and now we
need a subset of substitutions to occur even on _expr, but for Dynamo
output graph deferred runtime assert, this cannot have happened yet, so
I think this completely solves the problem in this neck of the woods.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: e1e815cb42af508a65fd80caeabfa3888ba5e82f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121079