[DCP] RFC, first pass at a DCP logger
This diff represents a RFC on the overall approach of improving DCP observability and debug ability internally. Any lingering TODO's, tests, or general code cleanliness will be taken care of before we merge.

A large motivator for this work is unification with model store components. To that extent, please note any existing model store logs should not be changing. In the "Unified Checkpointer", existing model store logs should be be migrated either to one of the plugins passed to DCP API's (StorageReader/Writer, Save/LoadPlanner), or to the trainer level.

Differential Revision: [D54512626](https://our.internmc.facebook.com/intern/diff/D54512626/)

ghstack-source-id: 217637574
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121352